gcc-4 defaults forced me to edit many many files to get rid of the char * vs. unsigned char * warnings.  The GNU coding standards used to say that these mismatches are okay and better than a bunch of casts. Obviously this has changed now.

+++ tools/gpg-connect-agent.c
@@ -458,7 +458,7 @@ static int
 read_and_print_response (assuan_context_t ctx)
 {
   char *line;
-  int linelen;
+  size_t linelen;
   assuan_error_t rc;
   int i, j;
